Move unsegmented text handling to new module.
